post create done

+++ apps/panel/src/global/Models/videoModel.tsx
@@ -9,6 +9,14 @@ import routes from 'global/Constants/routes';
 const title = 'ویدیو';
 const englishTitle = 'video';
 
+const inputs = (
+  <>
+    <TcFormItem label='عنوان' name='title'>
+      <TcInput placeholder='عنوان' />
+    </TcFormItem>
+  </>
+);
+
 const filterInputs = (
   <>
     <TcFormItem name='title' label='عنوان'>
@@ -17,47 +25,53 @@ const filterInputs = (
   </>
 );
 
-const columns = (handleDelete: (_id: string) => void) => [
-  {
-    title: 'ویدیو',
-    key: 'video',
-    dataIndex: 'video',
-    width: 170,
-    render: (_text: string, record: IVideoRead) => <VideoItem video={record} size='small' />,
-  },
-  {
-    title: 'عنوان SEO',
-    key: 'title',
-    dataIndex: 'title',
-  },
-  {
-    title: 'کیفیت ها',
-    key: 'qualities',
-    dataIndex: 'qualities',
-    render: (_text: string, record: IVideoRead) => (
-      <p className='flex gap-2'>
-        {record.qualityVariations.map((vid) => (
-          <span key={vid.size}>{vid.size}</span>
-        ))}
-      </p>
-    ),
-  },
-  {
-    title: 'عملیات',
-    key: 'operation',
-    render: (_text: any, record: any) => (
-      <div className='flex'>
-        {/* <TcDetailIcon to={routes.effectivePointTypesDetail.path + '/' + record.id} /> */}
-        <TcEditIcon to={routes.videoEdit.path + '/' + record._id} />
-        <TcDeleteIcon onConfirm={() => handleDelete(record._id)} />
-      </div>
-    ),
-  },
-];
+const columns = (handleDelete?: (_id: string) => void) => {
+  const columns: any = [
+    {
+      title: 'ویدیو',
+      key: 'video',
+      dataIndex: 'video',
+      width: 170,
+      render: (_text: string, record: IVideoRead) => <VideoItem video={record} size='small' />,
+    },
+    {
+      title: 'عنوان SEO',
+      key: 'title',
+      dataIndex: 'title',
+    },
+    {
+      title: 'کیفیت ها',
+      key: 'qualities',
+      dataIndex: 'qualities',
+      render: (_text: string, record: IVideoRead) => (
+        <p className='flex gap-2'>
+          {record.qualityVariations.map((vid) => (
+            <span key={vid.size}>{vid.size}</span>
+          ))}
+        </p>
+      ),
+    },
+  ];
+  if (handleDelete)
+    columns.push({
+      title: 'عملیات',
+      key: 'operation',
+      render: (_text: any, record: any) => (
+        <div className='flex'>
+          {/* <TcDetailIcon to={routes.effectivePointTypesDetail.path + '/' + record.id} /> */}
+          <TcEditIcon to={routes.videoEdit.path + '/' + record._id} />
+          <TcDeleteIcon onConfirm={() => handleDelete(record._id)} />
+        </div>
+      ),
+    });
+
+  return columns;
+};
 
 const videoModel = {
   title,
   englishTitle,
+  inputs,
   filterInputs,
   columns,
 };
